add rest extension definition
