+++ app.ts
@@ -3,7 +3,7 @@ import { engine } from '@kites/core';
 async function bootstrap() {
   // load config & autodiscover
   const app = await engine(true).init();
-  app.logger.info('Application started!');
+  app.logger.info('Kites application started!');
 }
 
 bootstrap();
