+++ packages/rest/index.ts
@@ -1,6 +1,8 @@
 import 'reflect-metadata';
 
-import { ExtensionOptions, IKites, KitesExtension } from '@kites/core';
+import { ExtensionOptions, KitesExtension } from '@kites/core';
+import config from './kites.config';
+import main from './main';
 import { RestExtension } from './rest.extension';
 
 export * from './decorators';
@@ -14,16 +16,10 @@ export {
   RestExtension
 };
 
-export default function Rest(kites: IKites, definition: KitesExtension) {
-  kites.options.apiPrefix = kites.options.apiPrefix || '/api';
-
-  // if (kites.options.apiPrefix.substr(-1) !== '/') {
-  //   kites.options.apiPrefix += '/';
-  // }
-
-  definition.name = definition.name || 'Rest';
-  kites.options.rest = definition.options || {};
-
-  var extension = new RestExtension(kites, definition.options);
-  kites.initializeListeners.add(definition.name, extension.init.bind(extension, kites, definition.options));
+export default function (options?: ExtensionOptions) {
+  const definition: KitesExtension = config;
+  definition.options = options;
+  definition.main = main;
+  definition.directory = __dirname;
+  return definition;
 }
