+++ packages/rest/kites.config.ts
@@ -1,0 +1,8 @@
+export = {
+  name: 'rest',
+  main: 'main',
+  dependencies: [],
+  options: {
+    // discover: true
+  }
+};
